display notes acc to authed userid

+++ backend/src/controllers/notes.ts
@@ -2,12 +2,15 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertisDefined } from "../util/assertisDefined";
 
 // creating an endpoint for a HTTP get request
 export const getNotes : RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId;
 
     try {
-        const notes = await NoteModel.find().exec();
+        assertisDefined(authenticatedUserId);
+        const notes = await NoteModel.find({ userId: authenticatedUserId}).exec();
         res.status(200).json(notes);
     } catch (error) {
         next(error);
@@ -15,9 +18,12 @@ export const getNotes : RequestHandler = async (req, res, next) => {
 };
 
 export const getNote: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId;
+
     const noteId = req.params.noteId;
     try {
-        
+        assertisDefined(authenticatedUserId);
+
         if(!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid node ID.")
         }
@@ -25,6 +31,10 @@ export const getNote: RequestHandler = async (req, res, next) => {
 
         if (!note) {
             throw createHttpError(404, "Note not found.");
+        }
+
+        if(!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note!")
         }
 
         res.status(200).json(note);
@@ -42,13 +52,17 @@ interface CreateNoteBody {
 export const createNote: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = async (req, res, next) => {
     const title = req.body.title;
     const text = req.body.text;
+    const authenticatedUserId = req.session.userId;
     
     try {
+        assertisDefined(authenticatedUserId);
+
         if (!title) {
             throw createHttpError(400, "Note must have a title.");
         }
 
         const newNote = await NoteModel.create({
+            userId: authenticatedUserId,
             title : title,
             text : text,
         });
@@ -71,7 +85,11 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
     const noteId = req.params.noteId;
     const newTitle = req.body.title;
     const newText = req.body.text;
+    const authenticatedUserId = req.session.userId;
+
     try {
+        assertisDefined(authenticatedUserId);
+
         if(!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid node ID.")
         }
@@ -84,6 +102,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
         if (!note) {
             throw createHttpError(404, "Note not found.");
+        }
+
+        if(!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note!")
         }
 
         // update note
@@ -99,14 +121,21 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertisDefined(authenticatedUserId);
+
         if(!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid node ID.")
         }
         const note = await NoteModel.findById(noteId).exec();
         if(!note) {
             throw createHttpError(404, "Note not found.");
+        }
+
+        if(!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note!")
         }
 
         await note.deleteOne().then(() => {
